+++ src/ultimate_notion/database.py
@@ -4,18 +4,19 @@ from __future__ import annotations
 
 from typing import cast
 
-from pydantic import BaseModel
+from pydantic import BaseModel, ValidationError
 from typing_extensions import Self
 
 from ultimate_notion.blocks import ChildrenMixin, DataObject
 from ultimate_notion.core import get_active_session, get_repr
+from ultimate_notion.errors import ReadOnlyPropertyError, SchemaError
 from ultimate_notion.file import Emoji, FileInfo
 from ultimate_notion.obj_api import blocks as obj_blocks
 from ultimate_notion.obj_api import objects as objs
 from ultimate_notion.page import Page
 from ultimate_notion.query import Query
 from ultimate_notion.rich_text import Text, camel_case, snake_case
-from ultimate_notion.schema import Property, PropertyType, PropertyValue, ReadOnlyPropertyError, Schema, SchemaError
+from ultimate_notion.schema import Property, PropertyType, Schema
 from ultimate_notion.view import View
 
 
@@ -189,44 +190,36 @@ class Database(DataObject[obj_blocks.Database], wraps=obj_blocks.Database):
         msg = 'Use .is_empty instead of bool(db) to check if a database is empty.'
         raise RuntimeError(msg)
 
-    def pydantic_model(self) -> BaseModel:
-        """Return a Pydantic model for this database."""
-        # Check https://github.com/ultimate-notion/ultimate-notion/issues/32 for details
-        raise NotImplementedError
-
     def create_page(self, **kwargs) -> Page:
         """Create a page with properties according to the schema within the corresponding database."""
-
-        # ToDo: let pydantic_model check the kwargs and raise an error if something is wrong
-        #       this would also generate a nice error message for the user!
-        schema_kwargs = {prop.attr_name: prop for prop in self.schema.get_props()}
-        if not set(kwargs).issubset(set(schema_kwargs)):
-            add_kwargs = set(kwargs) - set(schema_kwargs)
+        attr_to_name = {prop.attr_name: prop.name for prop in self.schema.get_props()}
+        if not set(kwargs).issubset(set(attr_to_name)):
+            add_kwargs = set(kwargs) - set(attr_to_name)
             msg = f"kwargs {', '.join(add_kwargs)} not defined in schema"
             raise SchemaError(msg)
+        if ro_props := set(kwargs) & {prop.attr_name for prop in self.schema.get_ro_props()}:
+            msg = f"Read-only properties {', '.join(ro_props)} cannot be set"
+            raise ReadOnlyPropertyError(msg)
 
-        schema_dct = {}
-        for kwarg, value in kwargs.items():
-            prop = schema_kwargs[kwarg]
-            prop_value_cls = prop.type.prop_value  # map schema to page property
-            # ToDo: Check at that point in case of select option if the option is already defined in Schema!
-
-            if prop_value_cls.readonly:
-                raise ReadOnlyPropertyError(prop)
-
-            prop_value = value if isinstance(value, PropertyValue) else prop_value_cls(value)
-
-            schema_dct[schema_kwargs[kwarg].name] = prop_value.obj_ref
+        schema_kwargs = {attr_to_name[attr]: value for attr, value in kwargs.items()}
+        validator = self.schema.to_pydantic_model(with_ro_props=False)
+        try:
+            py_page: BaseModel = validator(**schema_kwargs)
+        except ValidationError as e:
+            msg = f'Invalid keyword arguments or overwriting read-only properties:\n{e}'
+            raise SchemaError(msg) from e
 
         session = get_active_session()
-        page_obj = session.api.pages.create(parent=self.obj_ref, properties=schema_dct)
+        page_obj = session.api.pages.create(
+            parent=self.obj_ref, properties=py_page.model_dump(mode='json', by_alias=True)
+        )
         page = Page.wrap_obj_ref(page_obj)
         session.cache[page.id] = page
         return page
 
     def to_markdown(self) -> str:
-        """Return the content of this database as Markdown."""
-        raise NotImplementedError
+        """Return the reference to this database as Markdown."""
+        return self._to_markdown()
 
     def _to_markdown(self) -> str:
         """Return the reference to this database as Markdown."""