+++ tests/test_view.py
@@ -5,10 +5,11 @@ from __future__ import annotations
 import pytest
 
 import ultimate_notion as uno
+from ultimate_notion.props import PropertyValue
 
 
 @pytest.mark.vcr()
-def test_select(contacts_db: uno.Database):
+def test_select(contacts_db: uno.Database) -> None:
     view = contacts_db.get_all_pages()
 
     sub_view = view.select('Name', 'Role')
@@ -22,7 +23,7 @@ def test_select(contacts_db: uno.Database):
 
 
 @pytest.mark.vcr()
-def test_rows(contacts_db: uno.Database):
+def test_rows(contacts_db: uno.Database) -> None:
     view = contacts_db.get_all_pages()
     rows = view.to_rows()
     assert len(rows) == len(view)
@@ -34,7 +35,7 @@ def test_rows(contacts_db: uno.Database):
 
 
 @pytest.mark.vcr()
-def test_index(contacts_db: uno.Database):
+def test_index(contacts_db: uno.Database) -> None:
     view = contacts_db.get_all_pages()
     assert not view.has_index
     view = view.with_index('my_index')
@@ -49,7 +50,7 @@ def test_index(contacts_db: uno.Database):
 
 
 @pytest.mark.vcr()
-def test_clone(contacts_db: uno.Database):
+def test_clone(contacts_db: uno.Database) -> None:
     view = contacts_db.get_all_pages()
     assert len(view) == 10
     short_view = view.limit(3)
@@ -58,7 +59,7 @@ def test_clone(contacts_db: uno.Database):
 
 
 @pytest.mark.vcr()
-def test_reverse(contacts_db: uno.Database):
+def test_reverse(contacts_db: uno.Database) -> None:
     short_view = contacts_db.get_all_pages().limit(3)
     row_0 = short_view.get_row(0)
     row_2 = short_view.get_row(2)
@@ -68,7 +69,20 @@ def test_reverse(contacts_db: uno.Database):
 
 
 @pytest.mark.vcr()
-def test_to_pandas(task_db: uno.Database):
+def test_to_pandas(task_db: uno.Database) -> None:
     view = task_db.get_all_pages()
     df = view.to_pandas()
     assert len(view) == len(df)
+
+
+@pytest.mark.vcr()
+def test_to_pydantic(task_db: uno.Database) -> None:
+    view = task_db.get_all_pages()
+    tasks = view.to_pydantic()
+    assert len(view) == len(tasks)
+
+    task = tasks[0]
+    for field_attr in task.model_fields:
+        field = getattr(task, field_attr)
+        assert isinstance(field, PropertyValue)
+        _ = field.value